Function violet::engine::make_html

pub fn make_html(deps: Vec<Event>) -> Result<String, Error>

Generates a HTML representation from a Vec MOCK IMPLEMENTATION

Will need to have knowledge of our defined schema to generate the event-embedded HTML.

The algorithm relies on the HTML schema to be built as components, i.e. rowcol components, and header/footer components, and works as so:

A workaround to not knowing how many events we’ll have to handle at runtime is to simply loop through them and generate a rowcol component for each. This is done by simply mapping each event’s fields to the columns we’ve already defined for our statements

Header/footer components are then appended to the rowcol components.

The algorithm returns a fully-formed HTML statement as Ok(String)